pyproject.toml
[tool.poetry]
name = "posts and comments"
version = "0.1.0"
description = ""
authors = ["AntonSkyrda <mr.skyrda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.2"
python-dotenv = "^1.0.1"
django-simple-captcha = "^0.6.0"
psycopg2-binary = "^2.9.10"
bleach = "^6.1.0"
lxml = "^5.3.0"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
django-ckeditor-5 = "^0.2.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
